Create file for ResultStream module and fix the tests for new format of response from Profile.one and Profile.all
